This is a patch from <dougb@dougbarton.us> with minor adjustments from <phil@ipom.com>
Doug pointed out that his version of gpg2 won't auto-create a keyring that doesn't exist and shortly there after someone with 1.4.10 reported the same issue. I cannot reproduce this on either 1.4.11 or 2.0.18, but touching the file will increase compatibility.

Doug's patch added the line of the code, all I did was rename the function to be more clear about it's new functionality and add a comment.

Signed-off-by: Phil Dibowitz <phil@ipom.com>