Better Handling of UID Indexes
tldr; we now use the index that gpg gives us instead of trying to
determine it.

Previously we kept track of which "index" mapped to a UID by the order
the appeared in the output. This generall works, unless there's a UID
type we don't support (like photos), in which case our index get off,
and we end up trying to sign the wrong thing.

Now we pull the index from the output and store it in the object. This
makes the code much more resilient as well as simplifying the code
quite a bit.

This closes bug #13.

Signed-off-by: Phil Dibowitz <phil@ipom.com>