Add KEY_CONSIDERED 1 more place; normalize
Erik (@eskultety on GitHub) reports in PR #44 that the `KEY_CONSIDERED`
handling is missing in one place.  This is the change in 687f650 written
by @civodul (which was sufficient for my use-case); while looking, I saw
that the quoting and parentheses style from that patch was inconsistent
with surrounding code.

This patch adds the missing case and normalizes style of the other
changes from 687f650