Fix buggy statement re template text
The statement about the templatization of the email said `%(foo)` where
it should have said `%(foo)s` per normal Python percent interpolation.

Fix the statement to be correct, and for the normalization, for each
`foo` expand to `%(foo)s` instead of `[FOO]` so that the output can be
copy/pasted into a file and used as an intact starting point.