Work with GnuPG 2.1 and --policy-url option (#44)
* Handle gpg output terminating unexpectedly

Avoid going into an infinite loop reading on a pipe which returns empty
strings forever at EOF.

* Handle KEY_CONSIDERED gpg response

This commits the patch from GitHub user `civodul` provided in issue #39

This closes #39

* Work with policy URLs once more

Keyids must come at the end of the command-line.  Options are not
handled sanely after keyids.  So move the policy-URL option to earlier
in the invoked command-line when asking gpg to sign the key.

* Add KEY_CONSIDERED 1 more place; normalize

Erik (@eskultety on GitHub) reports in PR #44 that the `KEY_CONSIDERED`
handling is missing in one place.  This is the change in 687f650 written
by @civodul (which was sufficient for my use-case); while looking, I saw
that the quoting and parentheses style from that patch was inconsistent
with surrounding code.

This patch adds the missing case and normalizes style of the other
changes from 687f650