[pius-keyring-mgr] Always provide common format keyrings
pius-keyring-mgr should always provide a common format of keyring. Newer
gpg's will make a pbx format by default, but this isn't easily readable
by old versions, or other software.

To accomplish this, if we're creating a keyring for the first time, we
will create it off to the side, and then export it to the path the user
requested. From then on, we will just use that file.

Also, fix an error message.

Closes #80